// libs/modules/src/lib/enum.ts
export enum ModulesEnum {
  auth = "auth",
  dashboards = "dashboards",
  subscriptions = "subscriptions",
  profiles = "profiles",
  places = "places",
  members = "members",
  businessContacts = "business-contacts",
  notes = "notes",
  attachments = "attachments",
  schedule = "schedule",
  financial = "financial",
}